pyproject.toml
[project]
name = "vocca-medical-assistant"
version = "1.0.0"
description = "Vocca AI Medical Assistant for appointment booking"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # LiveKit Agents with all plugins
    "livekit-agents[deepgram,openai,silero,azure,turn-detector]>=1.2.1",
    "livekit-plugins-noise-cancellation>=0.2.5",
    
    # Web framework for availability API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    
    # Data validation and settings
    "pydantic>=2.11.0",
    "pydantic-settings>=2.6.0",
    
    # Utilities
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    # Formatting and linting
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.12.4",
    
    # Type checking
    "mypy>=1.17.0",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    
    # Development utilities
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hatch
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
